Move roc struct operations to a new "struct" module
